requirements.txt
langchain==0.3.27
langchain_community==0.3.27
langchain_core==0.3.72
langchain_google_genai==2.1.8
langgraph==0.6.2
python-dotenv==1.1.1
streamlit==1.47.1
